# backend/apidoc.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Um título de API",
    "description": "Descrição da API"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": {
    "Bearer": []
  },
...
  "paths": {
...
    "/profissionais": {
      "get": {
        "tags": ["PROFISSIONAIS"],
        "summary": "Retorna profissionais",
        "description": "Retorna todos profissionais existentes",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Operação bem sucedida"
          }
        }
      },
      "post": {
        "tags": ["PROFISSIONAIS"],
        "summary": "Adiciona profissional",
        "description": "Adiciona um profissional",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "profissional",
            "description": "Objeto que descreve um profissional",
            "schema": {
              "type": "object",
              "required": [
                "nome",
                "email",
                "senha",
                "fone",
                "especialidade",
                "descricao",
                "cro"
              ],
              "properties": {
                "nome": { "type": "string", "example": "Dr. João" },
                "senha": { "type": "string", "example": "12345" },
                "email": { "type": "string", "example": "drjoao@clinica.com" },
                "fone": { "type": "number", "example": "(11) 9876-3455" },
                "especialidade": {
                  "type": "string",
                  "example": "Endodontia"
                },
                "descricao": {
                  "type": "string",
                  "example": "Especialista em canais"
                },
                "cro": { "type": "string", "example": "56789-SP" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação Inserção bem sucedida"
          }
        }
      }
    },
    "/profissionais/{id}": {
      "put": {
        "tags": ["PROFISSIONAIS"],
        "summary": "Altera profissional",
        "description": "Altera um profissional",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "ID do profissional a ser alterada",
            "required": true
          },
          {
            "in": "body",
            "name": "profissional",
            "description": "Objeto que descreve um profissional",
            "schema": {
              "type": "object",
              "required": [
                "nome",
                "email",
                "senha",
                "fone",
                "especialidade",
                "descricao",
                "cro"
              ],
              "properties": {
                "nome": { "type": "string", "example": "Dr. João" },
                "senha": { "type": "string", "example": "12345" },
                "email": { "type": "string", "example": "drjoao@clinica.com" },
                "fone": { "type": "number", "example": "(11) 9876-3455" },
                "especialidade": {
                  "type": "string",
                  "example": "Endodontia"
                },
                "descricao": {
                  "type": "string",
                  "example": "Especialista em canais"
                },
                "cro": { "type": "string", "example": "56789-SP" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação Inserção bem sucedida"
          }
        }
      },
      "delete": {
        "tags": ["PROFISSIONAIS"],
        "summary": "Remove profissional",
        "description": "Remove um profissional na lista de profissionais",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "ID do profissional a ser removido",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Operação Remoção bem sucedida"
          },
          "400": {
            "description": "PROFISSIONAL NÃO ENCONTRADO"
          }
        }
      }
    },
...
    "/procedimentos/{id}": {
      "put": {
        "tags": ["PROCEDIMENTOS"],
        "summary": "Altera procedimento",
        "description": "Altera um procedimento",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "ID do procedimento a ser alterada",
            "required": true
          },
          {
            "in": "body",
            "name": "procedimento",
            "description": "Objeto que descreve um procedimento",
            "schema": {
              "type": "object",
              "required": ["nome", "descricao", "idProfissionais"],
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Cirurgia do dente do siso"
                },
                "descricao": {
                  "type": "string",
                  "example": "Remoção do dente do siso"
                },
                "idProfissionais": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": ""
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação Inserção bem sucedida"
          }
        }
      },
      "delete": {
        "tags": ["PROCEDIMENTOS"],
        "summary": "Remove procedimento",
        "description": "Remove um procedimento na lista de procedimentos",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "ID do procedimento a ser removido",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Operação Remoção bem sucedida"
          },
          "400": {
            "description": "PROCEDIMENTO NÃO ENCONTRADO"
          }
        }
      }
    }
  }
}